회원가입, 아이디, 비밀번호 찾기, 카카오 API 연동 로그인, 아웃

+++ app/src/main/java/com/example/java_test/MainActivity.java
@@ -5,8 +5,12 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.app.Activity;
 import android.content.Intent;
 import android.content.SharedPreferences;
+import android.content.pm.PackageInfo;
+import android.content.pm.PackageManager;
+import android.content.pm.Signature;
 import android.os.Bundle;
 import android.util.AndroidRuntimeException;
+import android.util.Base64;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -17,10 +21,14 @@ import android.widget.SimpleAdapter;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.kakao.usermgmt.UserManagement;
+import com.kakao.usermgmt.callback.LogoutResponseCallback;
+
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.security.MessageDigest;
 import java.util.ArrayList;
 import java.util.HashMap;
 
@@ -40,8 +48,10 @@ public class MainActivity extends AppCompatActivity {
     Intent login_put_intent;
     Intent array_put_intent;
     String result;
+    String strNickname=null;
+    String strProfile=null;
+    String strEmail=null;
     String mJsonString;
-    ArrayList<HashMap<String, String>> mArrayList;
 
 
     @Override
@@ -60,21 +70,43 @@ public class MainActivity extends AppCompatActivity {
 
         logout = findViewById(R.id.logout_bt);
 
+        strNickname = login_put_intent.getStringExtra("name");
+        strProfile = login_put_intent.getStringExtra("profile");
         result = login_put_intent.getStringExtra("result_data");
         mJsonString = result;
-        showResult();
+
+        if(strNickname != null) {
+            main_name.setText(strNickname);
+            person_num.setText(strProfile);
+        }
+        else {
+            showResult();
+        }
 
 
 
         logout.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                startActivity(new Intent(getApplication(), LoginActivity.class));
-                SharedPreferences auto = getSharedPreferences("auto", Activity.MODE_PRIVATE);
-                SharedPreferences.Editor editor = auto.edit();
-                editor.clear();
-                editor.commit();
-                MainActivity.this.finish();
+                if(strNickname == null) {
+                    startActivity(new Intent(getApplication(), LoginActivity.class));
+                    SharedPreferences auto = getSharedPreferences("auto", Activity.MODE_PRIVATE);
+                    SharedPreferences.Editor editor = auto.edit();
+                    editor.clear();
+                    editor.commit();
+                    MainActivity.this.finish();
+                }
+                else {
+                    UserManagement.getInstance().requestLogout(new LogoutResponseCallback() {
+                        @Override
+                        public void onCompleteLogout() {
+                            Intent intent = new Intent(MainActivity.this, LoginActivity.class);
+                            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
+                            startActivity(intent);
+                            finish();
+                        }
+                    });
+                }
             }
         });
